refactor(wallet-lib): refactor getTransactionHistory.js
